payment/requirements.txt
redis
redis[hiredis]
stripe
pymongo[srv]
pymongo
flask
flask-cors
python-dotenv